fix root page ssr
